refactor(holdings): address Task 5 review — split bridge file, doc comments, test cleanup (#1101)
- Move AssetHolding.positionSelection to its own AssetHolding+PositionSelection.swift (one primary type per file)
- Doc-comment PositionSelection.kind and displayLabel
- Note assetHoldings is computed (folds per access) in its doc comment
- Add @Test display strings and drop shadowing aud/amount locals in favour of the suite helpers
