+++ Domain/Models/PositionSelection.swift
@@ -8,20 +8,10 @@ import Foundation
 struct PositionSelection: Sendable, Hashable, Identifiable {
   /// The row id — an `assetKey` for a crypto rollup, otherwise an instrument id.
   let id: String
+  /// The instrument kind of the selected row; drives the filter-chip badge.
   let kind: Instrument.Kind
+  /// Short label shown in the filter chip (e.g. "ETH").
   let displayLabel: String
   /// The per-chain instrument ids this selection covers (1+).
   let instrumentIds: [String]
-}
-
-/// Bridge from a display row to its selection value (defined here, alongside
-/// `PositionSelection`, so `AssetHolding` compiles without depending on it).
-extension AssetHolding {
-  var positionSelection: PositionSelection {
-    PositionSelection(
-      id: id,
-      kind: kind,
-      displayLabel: displayLabel,
-      instrumentIds: contributingInstrumentIds)
-  }
 }
